Fix cast from size_t to long

+++ MqMessageService/MqMessageService.cpp
@@ -413,7 +413,7 @@ namespace shape {
     {
       TRC_INFORMATION("Send to MQ: " << std::endl << MEM_HEX_CHAR(message.data(), message.size()));
 
-      unsigned long toWrite = message.size();
+      unsigned long toWrite = static_cast<unsigned long>(message.size());
       unsigned long written = 0;
       bool reconnect = false;
       bool fSuccess;
